Created constants in lib dir. Created dynamic import for schema modules
