+++ lib/constants.js
@@ -1,0 +1,8 @@
+const SCHEMAS = [
+  ['component', 'c'],
+  ['redux-component', 'rc'],
+  ['redux-slice', 'rs'],
+  ['redux-config', 'rconf'],
+];
+
+module.exports = { SCHEMAS };
